[Comments] useEffectHandleOutsideClick
